--- a/lttoolbox/compiler.h
+++ b/lttoolbox/compiler.h
@@ -78,6 +78,17 @@ private:
   UString current_paradigm;
 
   /**
+   * The number of top-level entries read in this section.
+   */
+  size_t n_section_entries = 0;
+
+  /**
+   * The maximum number of top-level entries per section.
+   * If 0, no limit.
+   */
+  size_t max_section_entries = 0;
+
+  /**
    * The dictionary section being compiled
    */
   UString current_section;
@@ -381,7 +392,12 @@ public:
   /**
    * Set whether to allow parallel minimisation jobs
    */
-  void setJobs(bool jobs = false);
+  void setJobs(bool jobs);
+
+  /**
+   * Set how many top-level entries to allow in a section before starting a new one automatically
+   */
+  void setMaxSectionEntries(size_t m);
 
   /**
    * Set verbose output