Number is overridable with env var `LT_MAX_SECTION_ENTRIES` 
(e.g. `export LT_MAX_SECTION_ENTRIES=5000`).

Fixes #134

Smaller sections also minimise faster, though after a certain point
the binary starts growing and total time increases again. Benching on
nob.dix where main section is about 260.000 entries, all columns are
with -j (4 cores) except the ones marked -j0:

```
| max entries | cpu secs | wall secs | cpu% | mem kb    | main sections | bin size | -j0 wall | -j0 mem   |
|-------------+----------+-----------+------+-----------+---------------+----------+----------+-----------|
|         500 |    21.62 |     10.87 |  203 |   699_300 |           496 | 4.2M     |    14.91 |   360_524 |
|        1000 |    20.47 |     09.58 |  218 |   658_172 |           248 | 3.7M     |    13.30 |   332_688 |
|        5000 |    23.02 |     09.35 |  250 |   652_964 |            50 | 2.9M     |    15.50 |   302_288 |
|       10000 |    25.10 |     09.98 |  254 |   665_204 |            25 | 2.7M     |    17.49 |   300_580 |
|       25000 |    28.85 |     12.96 |  225 |   699_964 |            10 | 2.5M     |    18.78 |   328_408 |
|       50000 |    29.64 |     13.72 |  219 |   678_776 |             5 | 2.4M     |    21.19 |   431_388 |
|      100000 |    27.95 |     16.99 |  167 |   896_884 |             3 | 2.4M     |    24.07 |   627_192 |
|      500000 |    26.74 |     27.86 |   97 | 1_105_992 |             1 | 2.4M     |    28.66 | 1_105_700 |
```
50k gives no increase in bin size while cutting time in half. 5k gives
best time while only a slight increase in bin size. Note also less
memory usage with more sections.

Running lt-proc on 10k corpus lines, the 5k entry bin takes 7.7s, 50k
takes 3.4s, while the original bin takes 2.9s. So we lose some in
processing speed, but lt-proc is far from being the bottleneck in
corpus processing.