@unhammer 2022-03-29
Yeah … though at least the way I would use it is as a pure compilation-speedup, not intended for changing the behaviour of the resulting bin (so not something you would hardcode in a Makefile). I think the 50k default seems like it works pretty well so I'm not sure people would actually change it unless they were very interested in experimenting with making things faster.

(Maybe we should wait with a cli switch until someone argues for a use case?)